% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{function1}
\alias{function1}
\title{A Function}
\usage{
function1(p1 = "")
}
\arguments{
\item{p1}{First parameter. Defaults to "".}
}
\description{
This function allows you to...
}
\examples{
function1()
}
\keyword{cats}

