% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lulc_data}
\alias{lulc_data}
\title{LULC data example}
\format{A data frame with 732494 rows and 16 variables}
\usage{
data(lulc_data)
}
\description{
This is a dataset of land use and land cover (LULC)
included as an example in the package dotrules.
}
\details{
The variables are as follow:

\itemize{
  \item POINTID unique identifier
  \item maxlike_1985 LULC type for 1985
  \item DFU_85 ???
  \item maxlike_1991 LULC type for 1991
  \item DFU_91 ???
  \item maxlike_2000 LULC type for 2000
  \item DFU_00 ???
  \item F00_3by3_W ???
  \item F00_3by3_R ???
  \item F00_3by3_B ???
  \item F00_3by3_A ???
  \item F00_3by3_U ???
  \item dis_to_rd_1 ???
  \item dis_to_drn_1 ???
  \item Slope slope
  \item maxlike_2015 LULC type for 2015
}
}
\keyword{datasets}
