% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderpredictors.R
\name{orderpredictors}
\alias{orderpredictors}
\title{Order predictors based on entropy}
\usage{
orderpredictors(data, x, y)
}
\arguments{
\item{data}{Dataframe}

\item{x}{vector of variable names of predictors}

\item{y}{response variable}
}
\value{
A vector of predictor names ordered by increasing entropy and corresponding entropy values
}
\description{
Sort columns by increasing entropy values
based on the factors of the response variable.
}
\details{
The mean entropy value is calculated for each variables
based on every outcome in the response variable.
Then this value is used for sorting the predictors.
}
\examples{
data('lulc_data')

x <- c('maxlike_1985','DFU_85','maxlike_1991','DFU_91','maxlike_2000','DFU_00',
      'F00_3by3_W','F00_3by3_R','F00_3by3_B','F00_3by3_A','F00_3by3_U')
y <- 'maxlike_2015'

orderedx <- orderpredictors(lulc_data, x, y)

lulc_model_orderedx <- dotrules(lulc_data,orderedx$x,y, sampsize =.1)
}
