% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotrules.plot.R
\name{plot.dotrules}
\alias{plot.dotrules}
\title{Plot results}
\usage{
\method{plot}{dotrules}(x, ...)
}
\arguments{
\item{x}{dotrules object}

\item{...}{other arguments}
}
\description{
Plot method for a dotrules object
}
