% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotrules.print.R
\name{print.dotrules}
\alias{print.dotrules}
\title{Print results}
\usage{
\method{print}{dotrules}(x, ...)
}
\arguments{
\item{x}{dotrules object}

\item{...}{other arguments}
}
\description{
Print method for a dotrules object.
Print confusion matrices of the overall accuracy,
the training sample accuracy, and the testing sample
accuracy.
}
